requirements.txt
# Core dependencies
flask==2.3.3
werkzeug==2.3.7
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
opencv-python==4.8.0.74
tensorflow==2.13.0
keras==2.13.0

# Data processing and visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.16.1

# API and web
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0

# Testing and development
pytest==7.4.0
black==23.7.0
flake8==6.1.0
pylint==2.17.5

# Utilities
tqdm==4.65.0
pillow==10.0.0
python-magic==0.4.27

torch==2.0.1
torchvision==0.15.2
scikit-learn==1.2.2
pandas==2.0.2
numpy==1.24.3
matplotlib==3.7.1
seaborn==0.12.2
flask==2.3.2
opencv-python==4.7.0.72
mediapipe==0.10.0
pillow==9.5.0
joblib==1.2.0
scipy==1.10.1
librosa==0.10.0.post2
nltk==3.8.1
transformers==4.29.2
flask-login==0.6.2
flask-wtf==1.1.1
pyyaml==6.0
dlib==19.24.1
imutils==0.5.4
h5py==3.8.0 